x/cardchain/keeper/msg_server_product_details_test.go
package keeper_test

import (
	"testing"

	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"

	"github.com/DecentralCardGame/cardchain/x/cardchain/types"
)

func TestProductDetailsMsgServerCreate(t *testing.T) {
	_, srv, ctx := setupMsgServer(t)
	wctx := sdk.UnwrapSDKContext(ctx)

	creator := "A"
	for i := 0; i < 5; i++ {
		resp, err := srv.CreateProductDetails(wctx, &types.MsgCreateProductDetails{Creator: creator})
		require.NoError(t, err)
		require.Equal(t, i, int(resp.Id))
	}
}

func TestProductDetailsMsgServerUpdate(t *testing.T) {
	creator := "A"

	tests := []struct {
		desc    string
		request *types.MsgUpdateProductDetails
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgUpdateProductDetails{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateProductDetails{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgUpdateProductDetails{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	}
	for _, tc := range tests {
		t.Run(tc.desc, func(t *testing.T) {
			_, srv, ctx := setupMsgServer(t)
			wctx := sdk.UnwrapSDKContext(ctx)

			_, err := srv.CreateProductDetails(wctx, &types.MsgCreateProductDetails{Creator: creator})
			require.NoError(t, err)

			_, err = srv.UpdateProductDetails(wctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}

func TestProductDetailsMsgServerDelete(t *testing.T) {
	creator := "A"

	tests := []struct {
		desc    string
		request *types.MsgDeleteProductDetails
		err     error
	}{
		{
			desc:    "Completed",
			request: &types.MsgDeleteProductDetails{Creator: creator},
		},
		{
			desc:    "Unauthorized",
			request: &types.MsgDeleteProductDetails{Creator: "B"},
			err:     sdkerrors.ErrUnauthorized,
		},
		{
			desc:    "KeyNotFound",
			request: &types.MsgDeleteProductDetails{Creator: creator, Id: 10},
			err:     sdkerrors.ErrKeyNotFound,
		},
	}
	for _, tc := range tests {
		t.Run(tc.desc, func(t *testing.T) {
			_, srv, ctx := setupMsgServer(t)
			wctx := sdk.UnwrapSDKContext(ctx)

			_, err := srv.CreateProductDetails(wctx, &types.MsgCreateProductDetails{Creator: creator})
			require.NoError(t, err)
			_, err = srv.DeleteProductDetails(wctx, tc.request)
			if tc.err != nil {
				require.ErrorIs(t, err, tc.err)
			} else {
				require.NoError(t, err)
			}
		})
	}
}
